LUCENE-2115: cutover contrib tests to use generics
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@887524 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/org/apache/lucene/index/TestIndexReaderCloneNorms.java
@@ -256,8 +256,8 @@ public class TestIndexReaderCloneNorms extends LuceneTestCase {
     // System.out.println("modifyNormsForF1 maxDoc: "+n);
     for (int i = 0; i < n; i += 3) { // modify for every third doc
       int k = (i * 3) % modifiedNorms.size();
-      float origNorm = ((Float) modifiedNorms.get(i)).floatValue();
-      float newNorm = ((Float) modifiedNorms.get(k)).floatValue();
+      float origNorm =  modifiedNorms.get(i).floatValue();
+      float newNorm =  modifiedNorms.get(k).floatValue();
       // System.out.println("Modifying: for "+i+" from "+origNorm+" to
       // "+newNorm);
       // System.out.println(" and: for "+k+" from "+newNorm+" to "+origNorm);
